+++ src/test/java/nextstep/ladder2/PointTest.java
@@ -6,8 +6,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.Objects;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
@@ -57,46 +55,4 @@ public class PointTest {
         assertThat(second.move()).isEqualTo(1);
     }
 
-    private static class Point {
-        private final int index;
-        private final Direction direction;
-
-        Point(int index, Direction direction) {
-            this.index = index;
-            this.direction = direction;
-        }
-
-        public static Point first(boolean right) {
-            return new Point(0, Direction.first(right));
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o)
-                return true;
-            if (o == null || getClass() != o.getClass())
-                return false;
-            Point point = (Point) o;
-            return index == point.index && direction.equals(point.direction);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(index, direction);
-        }
-
-        public int move() {
-            if (direction.isRight()) {
-                return index + 1;
-            }
-            if (direction.isLeft()) {
-                return index - 1;
-            }
-            return index;
-        }
-
-        public Point next(boolean right) {
-            return new Point(index + 1, direction.next(right));
-        }
-    }
 }
